refactor: remove internal routes, fix Apollo→Lead service, clean up API
- Remove Apollo service dependency, use Lead service for leads/stats
- Delete internal.ts, move 3 scheduler routes to campaigns.ts/runs.ts
- Fix route ordering bug (/all before /:id params)
- Rename /all → /list, /stats → /campaigns/stats
- Remove redundant activate/pause routes, use PATCH with status
- Add requireApiKey to run write routes (POST runs, PATCH run status)
- Normalize DB status to ongoing/stopped, migration 0009
- Status API accepts "activate"/"stop", maps to DB values
