+++ .env.example
@@ -15,8 +15,6 @@ CLERK_SECRET_KEY='sk_live_...'
 CAMPAIGN_SERVICE_API_KEY='your-campaign-service-api-key'
 
 # Downstream services (stats aggregation)
-APOLLO_SERVICE_URL='http://localhost:3003'
-APOLLO_SERVICE_API_KEY='your-apollo-api-key'
 EMAILGENERATION_SERVICE_URL='http://localhost:3004'
 EMAILGENERATION_SERVICE_API_KEY='your-emailgen-api-key'
 POSTMARK_SERVICE_URL='http://localhost:3006'
