+++ drizzle/0009_normalize_status.sql
@@ -1,0 +1,3 @@
+-- Normalize campaign status to only 'ongoing' or 'stopped'
+-- ongoing stays as-is, everything else (draft, paused, active, null) -> stopped
+UPDATE campaigns SET status = 'stopped' WHERE status IS NULL OR status NOT IN ('ongoing');
