+++ src/index.ts
@@ -13,8 +13,6 @@ const openapiPath = join(__dirname, "..", "openapi.json");
 import healthRoutes from "./routes/health.js";
 import campaignsRoutes from "./routes/campaigns.js";
 import runsRoutes from "./routes/runs.js";
-import internalRoutes from "./routes/internal.js";
-
 const app = express();
 const PORT = process.env.PORT || 3003;
 
@@ -35,7 +33,6 @@ app.use(express.json());
 app.use(healthRoutes);
 app.use(campaignsRoutes);
 app.use(runsRoutes);
-app.use("/internal", internalRoutes);
 
 // OpenAPI spec endpoint
 app.get("/openapi.json", (_req, res) => {
