+++ src/routes/runs.ts
@@ -2,12 +2,53 @@ import { Router } from "express";
 import { eq, and } from "drizzle-orm";
 import { db } from "../db/index.js";
 import { campaigns, orgs } from "../db/schema.js";
-import { clerkAuth, requireOrg, AuthenticatedRequest } from "../middleware/auth.js";
+import { clerkAuth, requireOrg, requireApiKey, AuthenticatedRequest } from "../middleware/auth.js";
 import { validateBody } from "../middleware/validate.js";
 import { ensureOrganization, listRuns, getRun, createRun, updateRun } from "@mcpfactory/runs-client";
 import { RunStatusUpdate } from "../schemas.js";
 
 const router = Router();
+
+// === Scheduler routes (API-key authed, must be before :runId routes) ===
+
+/**
+ * GET /campaigns/:campaignId/runs/list - Get campaign runs (for scheduler, no org scoping)
+ */
+router.get("/campaigns/:campaignId/runs/list", requireApiKey, async (req, res) => {
+  try {
+    const { campaignId } = req.params;
+
+    const campaign = await db.query.campaigns.findFirst({
+      where: eq(campaigns.id, campaignId),
+      columns: { orgId: true },
+    });
+    if (!campaign) {
+      return res.status(404).json({ error: "Campaign not found" });
+    }
+
+    const org = await db.query.orgs.findFirst({
+      where: eq(orgs.id, campaign.orgId),
+      columns: { clerkOrgId: true },
+    });
+    if (!org) {
+      return res.status(404).json({ error: "Organization not found" });
+    }
+
+    const runsOrgId = await ensureOrganization(org.clerkOrgId);
+    const result = await listRuns({
+      organizationId: runsOrgId,
+      serviceName: "campaign-service",
+      taskName: campaignId,
+    });
+
+    res.json({ runs: result.runs });
+  } catch (error) {
+    console.error("[Campaign Service] Get campaign runs error:", error);
+    res.status(500).json({ error: "Internal server error" });
+  }
+});
+
+// === User routes (Clerk JWT authed) ===
 
 /**
  * GET /campaigns/:campaignId/runs - List all runs for a campaign
@@ -16,7 +57,6 @@ router.get("/campaigns/:campaignId/runs", clerkAuth, requireOrg, async (req: Aut
   try {
     const { campaignId } = req.params;
 
-    // Verify campaign ownership
     const campaign = await db.query.campaigns.findFirst({
       where: and(
         eq(campaigns.id, campaignId),
@@ -28,7 +68,6 @@ router.get("/campaigns/:campaignId/runs", clerkAuth, requireOrg, async (req: Aut
       return res.status(404).json({ error: "Campaign not found" });
     }
 
-    // Fetch runs from runs-service
     const runsOrgId = await ensureOrganization(req.clerkOrgId!);
     const result = await listRuns({
       organizationId: runsOrgId,
@@ -50,7 +89,6 @@ router.get("/campaigns/:campaignId/runs/:runId", clerkAuth, requireOrg, async (r
   try {
     const { campaignId, runId } = req.params;
 
-    // Verify campaign ownership
     const campaign = await db.query.campaigns.findFirst({
       where: and(
         eq(campaigns.id, campaignId),
@@ -71,10 +109,12 @@ router.get("/campaigns/:campaignId/runs/:runId", clerkAuth, requireOrg, async (r
   }
 });
 
+// === Service routes (API-key authed) ===
+
 /**
- * POST /campaigns/:campaignId/runs - Trigger a new run (internal/service use)
+ * POST /campaigns/:campaignId/runs - Create a new run
  */
-router.post("/campaigns/:campaignId/runs", async (req, res) => {
+router.post("/campaigns/:campaignId/runs", requireApiKey, async (req, res) => {
   try {
     const { campaignId } = req.params;
 
@@ -86,7 +126,6 @@ router.post("/campaigns/:campaignId/runs", async (req, res) => {
       return res.status(404).json({ error: "Campaign not found" });
     }
 
-    // Look up clerkOrgId from the org
     const org = await db.query.orgs.findFirst({
       where: eq(orgs.id, campaign.orgId),
     });
@@ -110,9 +149,9 @@ router.post("/campaigns/:campaignId/runs", async (req, res) => {
 });
 
 /**
- * PATCH /runs/:runId - Update run status (internal/service use)
+ * PATCH /runs/:runId - Update run status
  */
-router.patch("/runs/:runId", validateBody(RunStatusUpdate), async (req, res) => {
+router.patch("/runs/:runId", requireApiKey, validateBody(RunStatusUpdate), async (req, res) => {
   try {
     const { runId } = req.params;
     const { status } = req.body;
